🩹 fixed : renamed file with typo in lot show page
