+++ src/components/pages/lot/show/lot-farmerProduce.tsx
